pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "botobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "boto3>=1.37.22",
    "prompt-toolkit>=3.0.50",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "typer>=0.15.2",
]

[project.scripts]
botobot = "botobot.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/botobot"]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
